Fully document and finish re-arranging methods in DecodingStream.
Internal buffer management was also removed from this class, as it was found to be easier to leave this to the derived classes in a case-by-case basis. I'm still considering leaving a field in the base class to make sure naming is uniform, but I'm currently weighing the pros and cons of this approach.